+++ sdk/communication/AzureCommunication/Source/CommunicationIdentifierSerializer.swift
@@ -40,13 +40,11 @@ public class CommunicationIdentifierSerializer {
         switch kind {
         case .communicationUser:
             return CommunicationUserIdentifier(identifier: id)
-        case .callingApplication:
-            return CallingApplicationIdentifier(identifier: id)
         case .phoneNumber:
             guard let phoneNumber = identifier.phoneNumber else {
                 throw AzureError.client("Can't serialize CommunicationIdentifierModel: phoneNumber is undefined.")
             }
-            return PhoneNumberIdentifier(phoneNumber: phoneNumber, id: id)
+            return PhoneNumberIdentifier(phoneNumber: phoneNumber, rawId: id)
         case .microsoftTeamsUser:
             guard let isAnonymous = identifier.isAnonymous else {
                 throw AzureError.client("Can't serialize CommunicationIdentifierModel: isAnonymous is undefined.")
@@ -57,7 +55,7 @@ public class CommunicationIdentifierSerializer {
             guard let cloud = identifier.cloud else {
                 throw AzureError.client("Can't serialize CommunicationIdentifierModel: cloud is undefined.")
             }
-            return MicrosoftTeamsUserIdentifier(userId: microsoftTeamsUserId, isAnonymous: isAnonymous, identifier:id, cloudEnvironment: try deserialize(model: cloud))
+            return MicrosoftTeamsUserIdentifier(userId: microsoftTeamsUserId, isAnonymous: isAnonymous, rawId: id, cloudEnvironment: try deserialize(model: cloud))
         default:
             return UnknownIdentifier(identifier: id)
         }
@@ -67,14 +65,12 @@ public class CommunicationIdentifierSerializer {
         switch identifier {
         case let userIdentifier as CommunicationUserIdentifier:
             return CommunicationIdentifierModel(kind: .communicationUser, id: userIdentifier.identifier)
-        case let callingApplicationIdentifier as CallingApplicationIdentifier:
-            return CommunicationIdentifierModel(kind: .callingApplication, id: callingApplicationIdentifier.identifier)
         case let phoneNumberIdentifier as PhoneNumberIdentifier:
-            return CommunicationIdentifierModel(kind: .phoneNumber, id: phoneNumberIdentifier.id, phoneNumber: phoneNumberIdentifier.phoneNumber)
+            return CommunicationIdentifierModel(kind: .phoneNumber, id: phoneNumberIdentifier.rawId, phoneNumber: phoneNumberIdentifier.phoneNumber)
         case let microsoftTeamUserIdentifier as MicrosoftTeamsUserIdentifier:
             return CommunicationIdentifierModel(
                 kind: .microsoftTeamsUser,
-                id: microsoftTeamUserIdentifier.identifier,
+                id: microsoftTeamUserIdentifier.rawId,
                 microsoftTeamsUserId: microsoftTeamUserIdentifier.userId,
                 isAnonymous: microsoftTeamUserIdentifier.isAnonymous,
                 cloud: try serialize(cloud: microsoftTeamUserIdentifier.cloudEnviroment)